requirements.txt
openai>=1.0.0
httpx>=0.27.0
arxiv>=2.1.0
transformers>=4.44.0
sentence-transformers>=3.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
numpy>=1.26.0
loguru>=0.7.0
uvicorn
matplotlib>=3.7.0
fastapi>=0.104.0
